buoi1/kafka/python/requirements.txt
confluent-kafka
fastapi[all]
python-dotenv
numpy
pandas
Faker